+++ scripts/post-build.ts
@@ -1,0 +1,5 @@
+import { removeSync } from 'fs-extra'
+
+removeSync('dist/index.umd.d.ts')
+removeSync('dist/index.umd.d.ts.map')
+removeSync('temp')