+++ src/isArray.ts
@@ -1,0 +1,18 @@
+/**
+ * Whatever argument is `Array` or not.
+ *
+ * @param val - input any value
+ * @returns The result of `Array.isArray(val)`
+ *
+ * @example
+ * ```ts
+ * isArray([]) // true
+ * isArray(['hello', 'world']) // true
+ * isArray({}) // false
+ * ```
+ *
+ * @public
+ */
+const isArray = (val: unknown): val is any[] => Array.isArray(val)
+
+export { isArray }