+++ src/types/index.ts
@@ -29,4 +29,21 @@ export type Primitive =
  * @public
  */
 export type Ord = string | number | bigint | boolean | Date
+
+/**
+ * Alias for Empty values
+ *
+ * @public
+ */
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type Empty = '' | [] | {}
+
+/**
+ * Alias for Falsy values
+ *
+ * @public
+ */
+export type Falsy = false | '' | 0 | null | undefined
+
 export type InferArray<T> = T extends (infer R)[] ? R : never
+export type Whitespace = ' ' | '\t' | '\n'